Moved nth-of-type styling to GridContainer
