+++ src/components/section.js
@@ -4,7 +4,7 @@ import styled from "styled-components";
 
 const SectionGrid = styled.div`
   display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(300px, 1fr));
+  grid-template-columns: repeat(auto-fill, minmax(300px, 400px));
   grid-gap: 2rem;
   margin-bottom: 4rem;
